test(tau-coding-agent): split auth_and_provider tests (#1381)
